Implement file downloads using the requests module
Originally, wanted to write the script using only the standard library
to prevent having any external dependencies. Additionally wanted the
script to pick up the local system's default version of Python,
regardless of whether it was Python 2 or 3. This was so that the script
could be effectively ran on almost any system without needing to modify
the system -- no need to install anything new. However, requests is the
de-facto standard for making HTTP requests in Python and Python 2 is
going to be EoL soon. (Addiitonally, my methods of dynamically having
the script work for both Python 2 and 3 were very, very poor :) I should
have used the six PyPI compabaility ibrary instead :)